refactor: TableGroup 도메인 분리
