+++ src/main/java/kitchenpos/tablegroup/domain/TableGroupRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.table.domain;
+package kitchenpos.tablegroup.domain;
 
-import kitchenpos.table.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
