+++ src/test/java/kitchenpos/tablegroup/ui/TableGroupRestControllerTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.ui;
+package kitchenpos.tablegroup.ui;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
@@ -12,10 +12,10 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.time.LocalDateTime;
 import java.util.List;
-import kitchenpos.table.application.TableGroupService;
 import kitchenpos.table.application.dto.GroupOrderTableRequest;
 import kitchenpos.table.application.dto.TableGroupResult;
 import kitchenpos.table.application.dto.TableGroupingRequest;
+import kitchenpos.tablegroup.application.TableGroupService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
